[TIM-835] Make DatabaseLog the entrance point of the package.
