+++ timbuctoo-instancev4/src/main/java/nl/knaw/huygens/timbuctoo/experimental/databaselog/DatabaseLog.java
@@ -1,117 +1,147 @@
 package nl.knaw.huygens.timbuctoo.experimental.databaselog;
 
+
 import com.fasterxml.jackson.databind.ObjectMapper;
+import com.google.common.collect.Lists;
+import nl.knaw.huygens.timbuctoo.experimental.databaselog.entry.LogEntryFactory;
 import nl.knaw.huygens.timbuctoo.model.Change;
+import nl.knaw.huygens.timbuctoo.server.GraphWrapper;
+import org.apache.tinkerpop.gremlin.neo4j.process.traversal.LabelP;
+import org.apache.tinkerpop.gremlin.process.traversal.dsl.graph.GraphTraversal;
+import org.apache.tinkerpop.gremlin.process.traversal.dsl.graph.__;
 import org.apache.tinkerpop.gremlin.structure.Edge;
-import org.apache.tinkerpop.gremlin.structure.Property;
+import org.apache.tinkerpop.gremlin.structure.Element;
+import org.apache.tinkerpop.gremlin.structure.T;
 import org.apache.tinkerpop.gremlin.structure.Vertex;
 import org.slf4j.Logger;
 import org.slf4j.LoggerFactory;
 
-import java.io.BufferedWriter;
 import java.io.IOException;
-import java.nio.charset.Charset;
-import java.nio.file.Files;
-import java.nio.file.Paths;
+import java.util.Comparator;
+import java.util.List;
 
+/**
+ * <p>
+ * Currently (as of 2016-06-13) the Timbuctoo database contains all versions of all entities and all versions of all
+ * relations as one interlinked graph. The old versions are filtered out while querying. We want to simplify the
+ * database model and the code that uses the database directly, without losing information. Therefore we're moving the
+ * history to a separate logOutput.
+ * </p>
+ * <p>
+ * This class creates that log from the current format.
+ * </p>
+ * <p>
+ * It sorts all vertices (i.e. versions of entities) as one big chronological list and then iterates over it, creating
+ * the log entries and calling their addToLog method. //FIXME an object that is immediately removed?
+ * </p>
+ * <p>
+ * A complicating factor is that many vertices and edges share the exact same modified time. So we don't know the exact
+ * order in which they were added. For vertices we do not to care about the exact order, however an edge between two
+ * vertices can only be added once those vertices exist (obviously).
+ * </p>
+ * <p>
+ * Our solution is: for a given modified-time we first add {@link LogEntry}s for all vertices and then for all edges.
+ * </p>
+ * <p>
+ * After creating the {@link LogEntry} objects it calls the {@link LogOutput} to write the {@link LogEntry}s. At this
+ * moment they are written as text to a file, we will later replace this with a {@link LogOutput} implementation that
+ * creates log vertices in the database.
+ * </p>
+ */
 public class DatabaseLog {
-
   public static final Logger LOG = LoggerFactory.getLogger(DatabaseLog.class);
+  private final GraphWrapper graphWrapper;
+  private final LogEntryFactory logEntryFactory;
+  private final LogOutput logOutput;
   private final ObjectMapper objectMapper;
-  private final BufferedWriter writer;
 
-  public DatabaseLog() {
-    objectMapper = new ObjectMapper();
-    try {
-      writer = Files.newBufferedWriter(Paths.get("dblog"), Charset.forName("UTF-8"));
-    } catch (IOException e) {
-      throw new RuntimeException(e);
+  public DatabaseLog(GraphWrapper graphWrapper) {
+    this(graphWrapper, new LogEntryFactory());
+  }
+
+  DatabaseLog(GraphWrapper graphWrapper, LogEntryFactory logEntryFactory) {
+    this(graphWrapper, logEntryFactory, new LogOutput());
+  }
+
+  DatabaseLog(GraphWrapper graphWrapper, LogEntryFactory logEntryFactory, LogOutput logOutput) {
+    this.graphWrapper = graphWrapper;
+    this.logEntryFactory = logEntryFactory;
+    this.logOutput = logOutput;
+    this.objectMapper = new ObjectMapper();
+  }
+
+  public void generate() {
+    long curTimestamp = 0L;
+
+    @SuppressWarnings("unchecked") // union warns about unchecked.
+      GraphTraversal<Vertex, ? extends Element> results = graphWrapper
+      .getGraph()
+      .traversal()
+      .V().has("modified")
+      .not(__.has(T.label, LabelP.of("searchresult"))) // ignore search results
+      .union(__.identity(), __.outE().has("modified"))
+      .dedup().by(__.valueMap("rev", "tim_id")) // add each version once
+      .order()
+      .by("modified", (Comparator<String>) (o1, o2) -> {
+        try {
+          long timeStamp1 = getTimestampFromChangeString(o1);
+          long timeStamp2 = getTimestampFromChangeString(o2);
+          return Long.compare(timeStamp1, timeStamp2);
+        } catch (IOException e) {
+          LOG.error("Cannot convert change", e);
+          LOG.error("Change 1 '{}'", o1);
+          LOG.error("Change 2 '{}'", o2);
+          return 0;
+        }
+      });
+
+    List<Edge> edges = Lists.newArrayList();
+    List<Vertex> vertices = Lists.newArrayList();
+    for (; results.hasNext(); ) {
+      Element element = results.next();
+      String modifiedString = element.value("modified");
+      try {
+        long modifiedTimestamp = getTimestampFromChangeString(modifiedString);
+        if (curTimestamp != modifiedTimestamp) {
+          vertices.forEach(vertex -> logEntryFactory.createForVertex(vertex).appendToLog(logOutput));
+          vertices.clear();
+          edges.forEach(edge -> logEntryFactory.createForEdge(edge).appendToLog(logOutput));
+          edges.clear();
+          curTimestamp = modifiedTimestamp;
+        }
+
+        if (element instanceof Vertex) {
+          vertices.add((Vertex) element);
+        } else if (element instanceof Edge) {
+          edges.add((Edge) element);
+        } else {
+          LOG.error("Element type {} not expected.", element.getClass());
+        }
+
+      } catch (IOException e) {
+        LOG.error("Change '{}' could nog be parsed.", modifiedString);
+      }
+    }
+    // add remaining
+    vertices.forEach(vertex -> logEntryFactory.createForVertex(vertex).appendToLog(logOutput));
+    vertices.clear();
+    edges.forEach(edge -> logEntryFactory.createForEdge(edge).appendToLog(logOutput));
+    edges.clear();
+  }
+
+  private long getTimestampFromChangeString(String changeString) throws IOException {
+    return objectMapper.readValue(changeString, Change.class).getTimeStamp();
+  }
+
+  private static class IdRevComparator implements Comparator<Element> {
+    @Override
+    public int compare(Element o1, Element o2) {
+      String id1 = o1.<String>property("tim_id").orElse("");
+      Integer rev1 = o1.<Integer>property("rev").orElse(0);
+      String id2 = o2.<String>property("tim_id").orElse("");
+      Integer rev2 = o2.<Integer>property("rev").orElse(0);
+
+      return id1.equals(id2) ? rev1.compareTo(rev2) : id1.compareTo(id2);
     }
   }
-
-  public void newVertex(Vertex vertex) {
-    String modifiedString = vertex.value("modified");
-
-    try {
-      Change modified = objectMapper.readValue(modifiedString, Change.class);
-      writeAndFlush(
-        String.format("%d - Vertex with tim_id '%s' created.%n", modified.getTimeStamp(), vertex.value("tim_id")));
-    } catch (IOException e) {
-      throw new RuntimeException(e);
-    }
-  }
-
-  public void updateVertex(Vertex vertex) {
-    String modifiedString = vertex.value("modified");
-
-    try {
-      Change modified = objectMapper.readValue(modifiedString, Change.class);
-      writeAndFlush(
-        String.format("%d - Vertex with tim_id '%s' updated.%n", modified.getTimeStamp(), vertex.value("tim_id")));
-    } catch (IOException e) {
-      throw new RuntimeException(e);
-    }
-  }
-
-  public void newEdge(Edge edge) {
-    String modifiedString = edge.value("modified");
-    try {
-      Change modified = objectMapper.readValue(modifiedString, Change.class);
-      writeAndFlush(
-        String.format(
-          "%d - Edge with tim_id '%s' between Vertex with tim_id '%s' and Vertex with tim_id '%s' created.%n",
-          modified.getTimeStamp(),
-          edge.value("tim_id"),
-          edge.outVertex().value("tim_id"),
-          edge.inVertex().value("tim_id")));
-    } catch (IOException e) {
-      throw new RuntimeException(e);
-    }
-  }
-
-  public void updateEdge(Edge edge) {
-    String modifiedString = edge.value("modified");
-    try {
-      Change modified = objectMapper.readValue(modifiedString, Change.class);
-      writeAndFlush(String.format(
-        "%d - Edge with tim_id '%s' between Vertex with tim_id '%s' and Vertex with tim_id '%s' updated.%n",
-        modified.getTimeStamp(),
-        edge.value("tim_id"),
-        edge.outVertex().value("tim_id"),
-        edge.inVertex().value("tim_id")));
-    } catch (IOException e) {
-      throw new RuntimeException(e);
-    }
-  }
-
-  public void newProperty(Property property) {
-    try {
-      writeAndFlush(String.format("Property '%s' set to '%s'.%n", property.key(), property.value()));
-    } catch (IOException e) {
-      throw new RuntimeException(e);
-    }
-  }
-
-  public void updateProperty(Property property) {
-    try {
-      writeAndFlush(String.format("Property '%s' set to '%s'.%n", property.key(), property.value()));
-    } catch (IOException e) {
-      throw new RuntimeException(e);
-    }
-  }
-
-  public void deleteProperty(String propertyName) {
-
-    try {
-      writeAndFlush(String.format("Property '%s' removed.%n", propertyName));
-    } catch (IOException e) {
-      throw new RuntimeException(e);
-    }
-  }
-
-  private void writeAndFlush(String format) throws IOException {
-    writer.write(format);
-    writer.flush();
-  }
-
 }
