+++ timbuctoo-instancev4/src/main/java/nl/knaw/huygens/timbuctoo/experimental/databaselog/entry/CreateEdgeLogEntry.java
@@ -1,6 +1,6 @@
 package nl.knaw.huygens.timbuctoo.experimental.databaselog.entry;
 
-import nl.knaw.huygens.timbuctoo.experimental.databaselog.DatabaseLog;
+import nl.knaw.huygens.timbuctoo.experimental.databaselog.LogOutput;
 import nl.knaw.huygens.timbuctoo.experimental.databaselog.LogEntry;
 import nl.knaw.huygens.timbuctoo.util.StreamIterator;
 import org.apache.tinkerpop.gremlin.structure.Edge;
@@ -21,7 +21,7 @@ class CreateEdgeLogEntry implements LogEntry {
   }
 
   @Override
-  public void appendToLog(DatabaseLog dbLog) {
+  public void appendToLog(LogOutput dbLog) {
     dbLog.newEdge(edge);
 
     StreamIterator.stream(edge.properties())
