+++ timbuctoo-instancev4/src/main/java/nl/knaw/huygens/timbuctoo/experimental/databaselog/entry/CreateVertexLogEntry.java
@@ -1,6 +1,6 @@
 package nl.knaw.huygens.timbuctoo.experimental.databaselog.entry;
 
-import nl.knaw.huygens.timbuctoo.experimental.databaselog.DatabaseLog;
+import nl.knaw.huygens.timbuctoo.experimental.databaselog.LogOutput;
 import nl.knaw.huygens.timbuctoo.experimental.databaselog.LogEntry;
 import nl.knaw.huygens.timbuctoo.util.StreamIterator;
 import org.apache.tinkerpop.gremlin.structure.Vertex;
@@ -27,7 +27,7 @@ class CreateVertexLogEntry implements LogEntry {
 
 
   @Override
-  public void appendToLog(DatabaseLog dbLog) {
+  public void appendToLog(LogOutput dbLog) {
     dbLog.newVertex(vertex);
 
     StreamIterator.stream(vertex.properties())
