+++ timbuctoo-instancev4/src/main/java/nl/knaw/huygens/timbuctoo/experimental/databaselog/entry/UpdateEdgeLogEntry.java
@@ -1,6 +1,6 @@
 package nl.knaw.huygens.timbuctoo.experimental.databaselog.entry;
 
-import nl.knaw.huygens.timbuctoo.experimental.databaselog.DatabaseLog;
+import nl.knaw.huygens.timbuctoo.experimental.databaselog.LogOutput;
 import nl.knaw.huygens.timbuctoo.experimental.databaselog.LogEntry;
 import org.apache.tinkerpop.gremlin.structure.Edge;
 
@@ -23,7 +23,7 @@ class UpdateEdgeLogEntry implements LogEntry {
   }
 
   @Override
-  public void appendToLog(DatabaseLog dbLog) {
+  public void appendToLog(LogOutput dbLog) {
     dbLog.updateEdge(edge);
     propertyUpdater.updateProperties(dbLog);
   }
