+++ timbuctoo-instancev4/src/main/java/nl/knaw/huygens/timbuctoo/server/tasks/DbLogCreatorTask.java
@@ -4,7 +4,7 @@ import com.fasterxml.jackson.databind.ObjectMapper;
 import com.google.common.base.Stopwatch;
 import com.google.common.collect.ImmutableMultimap;
 import io.dropwizard.servlets.tasks.Task;
-import nl.knaw.huygens.timbuctoo.experimental.databaselog.DatabaseLogGenerator;
+import nl.knaw.huygens.timbuctoo.experimental.databaselog.DatabaseLog;
 import nl.knaw.huygens.timbuctoo.server.TinkerpopGraphManager;
 import org.slf4j.LoggerFactory;
 
@@ -13,13 +13,13 @@ import java.io.PrintWriter;
 public class DbLogCreatorTask extends Task {
   private final TinkerpopGraphManager graphManager;
   private final ObjectMapper objectMapper;
-  private final DatabaseLogGenerator logGenerator;
+  private final DatabaseLog logGenerator;
 
   public DbLogCreatorTask(TinkerpopGraphManager graphManager) {
     super("createlog");
     this.graphManager = graphManager;
     objectMapper = new ObjectMapper();
-    logGenerator = new DatabaseLogGenerator(graphManager);
+    logGenerator = new DatabaseLog(graphManager);
   }
 
   @Override
